split legacy graph gated dependency tests
